+++ CHANGES.md
@@ -6,6 +6,7 @@ Unreleased
   Options `-j-std` and `-std-json` are still available as backward-compatibility no-ops unless
   environment variable `ATDGEN_FAIL_DEPRECATED_OPTIONS` is set to `true` in
   which case their use results in an exception.
+* atd-parser: improve (syntax) error messages (#426)
 
 2.16.0 (2025-01-22)
 -------------------